+++ logger/nulllogger.go
@@ -13,6 +13,7 @@ func (n nullLogger) ReplaceField(field string, val interface{}) Logger { return 
 func (n nullLogger) WithField(field string, val interface{}) Logger    { return n }
 func (n nullLogger) WithFields(fields Fields) Logger                   { return n }
 func (n nullLogger) WithError(err error) Logger                        { return n }
+func (nullLogger) Log(level Level, msg string)                         {}
 func (nullLogger) Debug(msg string)                                    {}
 func (nullLogger) Info(msg string)                                     {}
 func (nullLogger) Warn(msg string)                                     {}